requirements.txt
asgiref==3.7.2
captcha==0.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
CurrencyConverter==0.17.15
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.60.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-multiselectfield==0.1.12
django-recaptcha==4.0.0
django-resized==1.0.2
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.6
whitenoise==6.6.0
